backend/requirements.txt
alembic==1.13.1
argon2-cffi==23.1.0
click==8.1.7
exif==1.6.0
fastapi==0.110.0
pillow==10.2.0
pydantic-settings==2.2.1
python-multipart==0.0.9
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
sqlmodel==0.0.16
